+++ derive_builder/CHANGELOG.md
@@ -10,6 +10,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Updated to `quote` 0.5.2 #120
 - Removed support for deprecated attributes #120
 
+### Changed
+- Using `#[builder(private)]` at the struct level will now emit a private builder. #99
+
 ### Added
 - Use `build_fn(private)` to generate a build method only accessible within the mod scope #89